cambios en oee por turno y catalogo de turnos

+++ database/migrations/2020_05_01_211231_create_oee_table.php
@@ -22,7 +22,7 @@ class CreateOeeTable extends Migration
             $table->integer('runTime')->nullable();
             $table->integer('availableTime')->nullable();
             $table->float('performance')->nullable();
-            $table->integer('ict')->nullable();
+            $table->float('ict')->nullable();
             $table->integer('totalPieces')->nullable();
             $table->float('quality')->nullable();
             $table->integer('goodParts')->nullable();
